requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
billiard==4.2.0
celery==5.4.0
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.110.2
flake8==7.0.0
flower==2.0.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
humanize==4.9.0
idna==3.7
isort==5.13.2
kombu==5.3.7
Mako==1.3.3
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==24.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pyasn1==0.6.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
redis==5.0.4
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.2.13
